refactor(plugin): clarify controller and handler naming
