+++ CruiserJumpPractice/CruiserJumpPractice.csproj
@@ -15,7 +15,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <!-- Runtime-provided dependencies -->
+    <!-- Compile-only game dependencies -->
     <PackageReference Include="BepInEx.Core" Version="5.4.21" IncludeAssets="compile" />
     <PackageReference Include="LethalCompany.GameLibs.Steam" Version="81.0.5-ngd.0" IncludeAssets="compile">
       <Aliases>LethalCompany</Aliases>
